cenario_python/requirements.txt
glfw
pygame
PyOpenGL
PyOpenGL-accelerate